feat: complete layout
