+++ src/core/base.ts
@@ -16,8 +16,9 @@ export const stateContext = createStateContext()
 
 export const DEFINE = {
   MODULE_ID: '\0stylex-dev',
-  MODULE_CSS: '@stylex-dev.css',
-  HIJACK_PLUGINS: ['vite:css', 'vite:css-post']
+  MODULE_CSS: '\0@stylex-dev.css',
+  HIJACK_PLUGINS: ['vite:css', 'vite:css-post'],
+  HIJACK_MIDDLEWARE: 'viteTransformMiddleware'
 }
 
 export function stylex(opts: StylexPluginOptions = {}): Plugin {
